mute [nfc]: Organize the muted-topics code as a model
Following the, er, model of unreadModel.
