+++ src/reduxTypes.js
@@ -15,7 +15,6 @@ import type { Action, DispatchableWithoutAccountAction } from './actionTypes';
 import type {
   Topic,
   Message,
-  MutedTopicTuple,
   CrossRealmBot,
   RealmEmojiById,
   RealmFilter,
@@ -32,9 +31,11 @@ import type {
   GlobalSessionState,
   SessionState,
 } from './session/sessionReducer';
+import type { MuteState } from './mute/muteModelTypes';
 import type { PmConversationsState } from './pm-conversations/pmConversationsModel';
 import type { UnreadState } from './unread/unreadModelTypes';
 
+export type { MuteState } from './mute/muteModelTypes';
 export type * from './actionTypes';
 
 /**
@@ -177,12 +178,6 @@ export type MessagesState = Immutable.Map<number, Message>;
 export type MigrationsState = $ReadOnly<{|
   version?: number,
 |}>;
-
-// TODO(#3918): Use stream IDs for muted topics, not stream names.
-//   Server issue for fixing that in the API: https://github.com/zulip/zulip/issues/21015
-//   Meanwhile, even while the server is still sending stream names,
-//   we should convert to stream IDs at the edge.
-export type MuteState = $ReadOnlyArray<MutedTopicTuple>;
 
 /** A map from user IDs to the Unix timestamp in seconds when they were muted. */
 export type MutedUsersState = Immutable.Map<UserId, number>;
